move cmdfx to internal
