+++ cmd/kbbi/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 
-	"github.com/raf555/kbbi-api/cmd/cmdfx"
+	"github.com/raf555/kbbi-api/cmd/internal/cmdfx"
 	"github.com/raf555/kbbi-api/internal/dictionary/dictionaryfx"
 	"github.com/raf555/kbbi-api/internal/home/homefx"
 	httpfx "github.com/raf555/kbbi-api/internal/http/fx"
